+++ src/main/kotlin/xyz/theprogramsrc/simplecoreapi/global/exceptions/ModuleDisableException.kt
@@ -1,0 +1,6 @@
+package xyz.theprogramsrc.simplecoreapi.global.exceptions
+
+class ModuleDisableException: RuntimeException {
+    constructor(message: String): super(message)
+    constructor(message: String, cause: Throwable): super(message, cause)
+}